rfctr: extract matrix and vector modules
* Use "matrix" instead of "slice" for objects that contain row and
  column vectors and are used by the _Assembler object. We already use
  the term "slice" for the interface object and it's unnecessarily
  confusing to overload that name for a subordinate object.

* Extract vector objects into vector module and matrix objects into
  matrix module, leaving core frozen-slice objects in slices module.
